fix product resources

+++ project/app/DTO/CartDTO.php
@@ -11,13 +11,13 @@ readonly class CartDTO
     /**
      * @param User $user
      * @param Product[] $products
-     * @param int $cost
+     * @param float $cost
      * @param object{category: object{id: int, name: string, limit: int}, in_stock: int, out_stock: int, total: int}[] $count_info
      */
     public function __construct(
         public User             $user,
         public array|Collection $products,
-        public int              $cost,
+        public float              $cost,
         public array            $count_info
     )
     {
